+++ src/entrypoint/mod.rs
@@ -7,27 +7,31 @@ use std::{
 use version_compare::Version;
 
 pub fn run() {
-    // Generate the lazymc config file
-    config::generate();
-
-    // Create a new Command to run "lazymc start"
-    info!(target: "lazymc-docker-proxy::entrypoint", "Starting lazymc process...");
-
-    // true if PUBLIC_VERSION env var is set and is less than 1.20.3
+    // lazymc dropped support minecraft servers with version less than 1.20.3
     let is_legacy: bool = match var("PUBLIC_VERSION") {
         Ok(version) => Version::from(version.as_ref()) < Version::from("1.20.3"),
         Err(_) => false,
     };
 
+    // Generate the lazymc config file
+    config::generate(is_legacy);
+
+    // Create a new Command to run "lazymc start"
+    info!(target: "lazymc-docker-proxy::entrypoint", "Starting lazymc process...");
+
     // if is_legacy is true, run lazymc-legacy instead of lazymc
-    match is_legacy {
-        true => debug!(target: "lazymc-docker-proxy::entrypoint", "Running legacy version of lazymc as server protocol version is less than 1.20.3"),
-        false => debug!(target: "lazymc-docker-proxy::entrypoint", "Running newer version of lazymc as server protocol version is 1.20.3 or higher"),
-    }
-    let mut child: process::Child = Command::new(match is_legacy {
-        true => "lazymc-legacy",
-        false => "lazymc",
-    })
+    let program: &str = match is_legacy {
+        true => {
+            debug!(target: "lazymc-docker-proxy::entrypoint", "Running legacy version of lazymc as server protocol version is less than 1.20.3");
+            "lazymc-legacy"
+        },
+        false => {
+            debug!(target: "lazymc-docker-proxy::entrypoint", "Running newer version of lazymc as server protocol version is 1.20.3 or higher");
+            "lazymc"
+        },
+    };
+
+    let mut child: process::Child = Command::new(program)
         .arg("start")
         .spawn()
         .unwrap_or_else(|err| {